Make alfresco-module-registry use posix paths
This attempts to fix windows style paths as they are added to
the module registry. It also performs this repair when reading
the registry.

Technically this is sufficient to address #152 going forward.
However, I still plan to review the use of path.baseline() and
path.join() across the code base to see if we can simply force
POSIX style paths in cases where that will matter.

fixes #152